Change the order of the image to bring it to top
I converted the divs into flexbox and changed the order of the icons in the divs so they end up in the top of the div.

I had the headings at the top for accessibility. This way screen readers will read the heading before the icon even though the icon comes before visually.