Add margin between p and buttons in 343px
I noticed that the elements in the divs are closely packed together when the screen width gets smaller than 343px. And also the div height is fixed at 475px. The elements will go out of the limits of the divs when the screen gets smaller.

To fix this issue I added a margin top to buttons to have some space between the paragraphs and the buttons.

And I changed the height of divs to fit content in 343px to make the divs grow when they can't fit the content.